# btf: Optimizing BTF parsing by merging `readTypes` and `inflateRawTypes`